Refactor: Implement Admin Middleware

+++ backend/routes/api.php
@@ -11,6 +11,6 @@ Route::middleware(['auth:sanctum'])->group(function () {
         return $request->user();
     });
 
-    Route::post('/videos', [VideoController::class, 'store']);
-    Route::delete('/videos/{video}', [VideoController::class, 'destroy']);
+    Route::post('/videos', [VideoController::class, 'store'])->middleware(\App\Http\Middleware\EnsureAdmin::class);
+    Route::delete('/videos/{video}', [VideoController::class, 'destroy'])->middleware(\App\Http\Middleware\EnsureAdmin::class);
 });
